Fix a bug which potentially returns wrong statsitc tuple from internal cache.
Altough Relation statistics cache has the key of relation id and it is
expected to hash and match it as oid, the cache control data was
created withOUT HASH_FUNCTION flag for hash_create(). This causes a
false match between oids having a intermediate 0x00 byte and identical
lower bytes (as for little endian). Returning the wrong statistics
cache can result in server crash, or at least getting unexpected
plans.