Change release number of spec files to 2.
The previous release was based on the funcion-version of
HeapTupleGetDatum() so they cannot be loaded on the minor versions
earlier than 9.1.14, 9.2.9 and 9.3.5 with link failure of
HeapTupleHeaderGetDatum.

I confirmed that the commit(*1) does not affect pg_dbms_stats and
decided to build rpm releases on 9.1.13, 9.2.8 and 9.3.4 after this.

There has been one bug after 1.3.5 was released but it is a bug in
error message of rarely used SQL function so no increment of version
number.

*1: d3c9f9c5b57d2a8a7ab065a6b942703854ec5e97 for 9.1
    315442c01156d4f6c0f766ec51da090a613d03b8 for 9.2
    edde59db1339397ec389d45e4c0d0f86064dad1a for 9.3
    621a99a666ba1a27b852dc5ddc0e1b224c388f53 on master branch