Allow dbms_stats.prepare_statstweak() to take multi-word type names.
The function ends up with an error being given type names consists of
more than a word. This commit makes the spaces between keywords to be
replaced with underscores. The function takes the parameter as regtype
so every white space consists of exactly one space. Therefore
considering only the single-space case is enough.