Properly set privilege information to VariableStatData
As the fix for CVE-2017-7484, VariableStatData is required to be set
acl_ok properly and the locked stats will not be in effect if not. As
the result, pg_dbms_stats 1.3.8 doesn't work for the minor versions
since 9.2.21, 9.3.17, 9.4.12, 9.5.7 and 9.6.3. PG10 behaves so from
the first version. This commit fixes that.

The binaries built for the fixed versions can be loaded and work fine
on all minor versions containing pre-fixed versions, but the binaries
built for the pre-fixed versions still has the problem

The regression test was missing whether locked stats is 'actually
applied or not. The test items for the case is added.

The version is updated to 1.3.9.