Fix calling parameter of SPI_execute_plan
Fixed a long-standing bug but having done no harm over all existing
versions. SPI_execute_plan's third paramter is char [] but it is
called with bool []. The bool value is false and it is different from
'n', which represents !is_null so it has been worked as expected as the
result.