Fix the wrong statistic after export and import.
'0' is always passed to the value of ReadArrayBinary argument 'typmod'.
As a result, incorrect value statistics are imported on somebdata types
after importing statistics. For example, in the timestamp data type,
the millisecond value is rounded off.
In this commit, dbms_stats_array_recv gets the value of 'typmod' and
passes it to ReadArrayBinary.